code/requirements.txt
pandas>=2.2.0
numpy>=2.0.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
